findAll and pagination

+++ evision/src/main/java/com/kh/evision/car/model/service/CarService.java
@@ -12,7 +12,7 @@ public interface CarService {
 	void saveCar(CarDTO car, List<MultipartFile> files);
 	
 	// 차량 목록 조회
-	List<CarDTO> findAll(Long pageNo);
+	List<CarDTO> findAll(int pageNo);
 	
 	// 차량 정보 수정
 	CarDTO updateCar(Long carNo, CarDTO car, List<MultipartFile> files);
